+++ unprompted/shortcodes/basic/set.py
@@ -14,7 +14,7 @@ class Shortcode():
 		elif ("_prepend" in pargs): self.Unprompted.shortcode_user_vars[pargs[0]] = content + self.Unprompted.shortcode_user_vars[pargs[0]]
 		else: self.Unprompted.shortcode_user_vars[pargs[0]] = content
 		
-		self.Unprompted.log(f"Setting {pargs[0]} to {content}")
+		self.Unprompted.log(f"Setting {pargs[0]} to {self.Unprompted.shortcode_user_vars[pargs[0]]}")
 
-		if ("_out" in pargs): return(content)
+		if ("_out" in pargs): return(self.Unprompted.shortcode_user_vars[pargs[0]])
 		else: return("")